Samples the generation time for given incubation period samples

This is done assuming the generation time distribution of each individual is given by a skew-normal distribution with a location parameter equal to their incubation period.

Usage

incubation_to_generation_time(incubation_period_samples, alpha)

Arguments

incubation_period_samples

a positive numeric vector: samples from the incubation period distribution

alpha

a numeric scalar: skew parameter of the skew-normal distribution

Value

a numeric vector of equal length to the vector input to incubation_period_samples

Examples

incubation_to_generation_time(
  incubation_period_samples = c(1, 2, 3, 4, 1),
  alpha = 2
)
#> [1] 3.580762 3.688164 4.930596 6.407438 3.870629