ringbp: Simulate and evaluate contact tracing scenarios

A package to simulate transmission of an infectious disease and associated efforts to trace conacts and reduce onwards transmission.

See also

Author

Maintainer: Joel Hellewell Joel.Hellewell@lshtm.ac.uk (ORCID)

Authors:

Sam Abbott contact@samabbott.co.uk (ORCID)

Amy Gimma amy.gimma@lshtm.ac.uk

Tim Lucas tim.lucas@leicester.ac.uk

Sebastian Funk sebastian.funk@lshtm.ac.uk

Adam Kucharski adam.kucharski@lshtm.ac.uk

Other contributors:

Hugo Gruson hugo.gruson@data.org (ORCID) [contributor]

Carl A. B. Pearson carl.ab.pearson@gmail.com (ORCID) [contributor]

Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [contributor]