Move forward one generation in the branching process

Move forward one generation in the branching process

outbreak_step(
  case_data = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  r0isolated = NULL,
  r0community = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  prop.ascertain = NULL,
  k = NULL,
  quarantine = NULL,
  r0subclin = NULL,
  disp.subclin = NULL
)

Arguments

case_data

data.table of cases in outbreak so far; initially generated by outbreak_setup

disp.iso

numeric dispersion parameter for isolated cases (must be >0)

disp.com

numeric dispersion parameter for non-isolated cases (must be >0)

r0isolated

numeric reproduction number for isolated cases (must be >0)

r0community

numeric reproduction number for non-isolated cases (must be >0)

prop.asym

proportion of cases that are completely asymptomatic.

incfn

function samples from incubation period; generated by dist_setup

delayfn

function samples from the onset-to-hospitalisation delay; generated by dist_setup

prop.ascertain

numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)

k

numeric skew parameter for sampling the serial interval from the incubation period

quarantine

logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset

r0subclin

numeric reproduction number for sub-clinical non-isolated cases (must be >0)

disp.subclin

numeric dispersion parameter for sub-clincial non-isolated cases (must be >0)

Author

Joel Hellewell

Examples

if (FALSE) { # \dontrun{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(1.651524, 4.287786)
# generate initial cases
case_data <- outbreak_setup(num.initial.cases = 5,incfn,delayfn,k=1.95,prop.asym=0)
# generate next generation of cases
case_data <- outbreak_step(case_data = case_data,
disp.iso = 1,
disp.com = 0.16,
r0isolated = 0,
r0subclin = 1.25,
disp.subclin = 0.16,
r0community = 2.5,
prop.asym = 0,
incfn = incfn,
delayfn = delayfn,
prop.ascertain = 0,
k = 1.95,
quarantine = FALSE)[[1]]
} # }