Sweep across parameters

Explore scenarios using gridding with sampling for parameters not in the grid. Parameters that are included in the grid are currently hard coded. Use the future package to control parallisation outside of the function.

parameter_sweep(scenarios = NULL, samples = 1, sim_fn = NULL)

Arguments

scenarios

a data.frame: containing all gridded scenarios - see the examples for the required structure. Defaults to NULL.

samples

a positive integer scalar: the number of samples to take. Defaults to 1.

sim_fn

a function: defaults to NULL. The vectorised model simulation function - see the examples for usage.

Value

A nested data.table containing the parameters for each scenario and a nested list of output from scenario_sim().

Examples

library(data.table)
# Put parameters that are grouped by disease into this data.table
scenarios <- data.table(
  expand.grid(
    delay_group = list(data.table(
      delay = c("SARS", "Wuhan"),
      onset_to_isolation = c(
        \(x) rweibull(n = x, shape = 1.651524, scale = 4.287786),
        \(x) rweibull(n = x, shape = 2.305172, scale = 9.483875)
      )
    )),
    k_group = list(data.table(
      theta = c("<1%", "15%"),
      k = c(1, 0.88)
    )),
    index_R0 = c(1.1, 1.5),
    prop.asym = c(0, 0.1),
    control_effectiveness = seq(0, 1, 0.25),
    num.initial.cases = c(5, 10),
    quarantine = FALSE
  )
)

list_cols <- grep("_group", colnames(scenarios), value = TRUE)
non_list_cols <- setdiff(colnames(scenarios), list_cols)

expanded_groups <- scenarios[, rbindlist(delay_group), by = c(non_list_cols)]
expanded_k <- scenarios[, rbindlist(k_group), by = c(non_list_cols)]

scenarios <- merge(
  expanded_groups, expanded_k, by = non_list_cols, allow.cartesian = TRUE
)
scenarios[, scenario :=  1:.N]
#> Key: <index_R0, prop.asym, control_effectiveness, num.initial.cases, quarantine>
#>      index_R0 prop.asym control_effectiveness num.initial.cases quarantine
#>         <num>     <num>                 <num>             <num>     <lgcl>
#>   1:      1.1       0.0                     0                 5      FALSE
#>   2:      1.1       0.0                     0                 5      FALSE
#>   3:      1.1       0.0                     0                 5      FALSE
#>   4:      1.1       0.0                     0                 5      FALSE
#>   5:      1.1       0.0                     0                10      FALSE
#>  ---                                                                      
#> 156:      1.5       0.1                     1                 5      FALSE
#> 157:      1.5       0.1                     1                10      FALSE
#> 158:      1.5       0.1                     1                10      FALSE
#> 159:      1.5       0.1                     1                10      FALSE
#> 160:      1.5       0.1                     1                10      FALSE
#>       delay onset_to_isolation  theta     k scenario
#>      <char>             <list> <char> <num>    <int>
#>   1:   SARS      <function[1]>    <1%  1.00        1
#>   2:   SARS      <function[1]>    15%  0.88        2
#>   3:  Wuhan      <function[1]>    <1%  1.00        3
#>   4:  Wuhan      <function[1]>    15%  0.88        4
#>   5:   SARS      <function[1]>    <1%  1.00        5
#>  ---                                                
#> 156:  Wuhan      <function[1]>    15%  0.88      156
#> 157:   SARS      <function[1]>    <1%  1.00      157
#> 158:   SARS      <function[1]>    15%  0.88      158
#> 159:  Wuhan      <function[1]>    <1%  1.00      159
#> 160:  Wuhan      <function[1]>    15%  0.88      160

incub <- \(x) rweibull(n = x, shape = 1.65, scale = 4.28)
scenarios[, incubation_period := rep(list(incub), .N)]
#> Key: <index_R0, prop.asym, control_effectiveness, num.initial.cases, quarantine>
#>      index_R0 prop.asym control_effectiveness num.initial.cases quarantine
#>         <num>     <num>                 <num>             <num>     <lgcl>
#>   1:      1.1       0.0                     0                 5      FALSE
#>   2:      1.1       0.0                     0                 5      FALSE
#>   3:      1.1       0.0                     0                 5      FALSE
#>   4:      1.1       0.0                     0                 5      FALSE
#>   5:      1.1       0.0                     0                10      FALSE
#>  ---                                                                      
#> 156:      1.5       0.1                     1                 5      FALSE
#> 157:      1.5       0.1                     1                10      FALSE
#> 158:      1.5       0.1                     1                10      FALSE
#> 159:      1.5       0.1                     1                10      FALSE
#> 160:      1.5       0.1                     1                10      FALSE
#>       delay onset_to_isolation  theta     k scenario incubation_period
#>      <char>             <list> <char> <num>    <int>            <list>
#>   1:   SARS      <function[1]>    <1%  1.00        1     <function[1]>
#>   2:   SARS      <function[1]>    15%  0.88        2     <function[1]>
#>   3:  Wuhan      <function[1]>    <1%  1.00        3     <function[1]>
#>   4:  Wuhan      <function[1]>    15%  0.88        4     <function[1]>
#>   5:   SARS      <function[1]>    <1%  1.00        5     <function[1]>
#>  ---                                                                  
#> 156:  Wuhan      <function[1]>    15%  0.88      156     <function[1]>
#> 157:   SARS      <function[1]>    <1%  1.00      157     <function[1]>
#> 158:   SARS      <function[1]>    15%  0.88      158     <function[1]>
#> 159:  Wuhan      <function[1]>    <1%  1.00      159     <function[1]>
#> 160:  Wuhan      <function[1]>    15%  0.88      160     <function[1]>

## Parameterise fixed paramters
sim_with_params <- purrr::partial(
  ringbp::scenario_sim,
  cap_max_days = 365,
  cap_cases = 5000,
  r0isolated = 0,
  disp.iso= 1,
  disp.subclin = 0.16,
  disp.com = 0.16
)

## parameter_sweep uses the future_lapply() function
## Default is to run sequntially on a single core
# future::plan("sequential")
## Set up multicore if using see ?future::plan for details
## Use the workers argument to control the number of cores used.
# future::plan("multiprocess")

## Run parameter sweep
sweep_results <- parameter_sweep(
  scenarios,
  sim_fn = sim_with_params,
  samples = 1
)

sweep_results
#>      scenario               data               sims
#>         <int>             <list>             <list>
#>   1:        1 <data.table[1x10]> <data.table[53x6]>
#>   2:       50 <data.table[1x10]> <data.table[53x6]>
#>   3:       88 <data.table[1x10]> <data.table[53x6]>
#>   4:       66 <data.table[1x10]> <data.table[53x6]>
#>   5:      145 <data.table[1x10]> <data.table[53x6]>
#>  ---                                               
#> 156:       38 <data.table[1x10]> <data.table[53x6]>
#> 157:      157 <data.table[1x10]> <data.table[53x6]>
#> 158:       11 <data.table[1x10]> <data.table[53x6]>
#> 159:      110 <data.table[1x10]> <data.table[53x6]>
#> 160:      118 <data.table[1x10]> <data.table[53x6]>